refactor(gui): move the asset-sync orchestration out of main.rs
SyncOutcome/AssetControl/run_asset_sync are a self-contained subsystem;
main.rs keeps startup wiring only.
