+++ crates/openlogi-gui/src/asset/sync.rs
@@ -10,11 +10,13 @@
 
 use std::fs;
 use std::path::Path;
+use std::time::Duration;
 
 use anyhow::{Context as _, Result};
+use backon::{BackoffBuilder, ExponentialBuilder};
 use openlogi_assets::http;
 use openlogi_assets::{BUTTONS_RENDER_FILES, DepotManifest, DeviceEntry, FetchOutcome};
-use openlogi_core::device::DeviceModelInfo;
+use openlogi_core::device::{DeviceInventory, DeviceModelInfo};
 use tracing::{debug, info, warn};
 
 /// Default origin for asset fetches. Overridable via `OPENLOGI_ASSETS`
@@ -192,4 +194,105 @@ fn pick_variant_filename(
     manifest
         .resource_for_variant(base_model_id, ext, resource_key)
         .map(str::to_string)
+}
+
+/// Result of one background asset-sync run, reported back to the select
+/// loop: whether the run succeeded, and which model keys it covered (folded
+/// into the synced set on success so the same device doesn't re-sync every
+/// snapshot).
+pub(crate) struct SyncOutcome {
+    pub(crate) ok: bool,
+    pub(crate) keys: Vec<String>,
+}
+
+/// Session-stable identity for a synced model: the HID++ model ids plus the
+/// extended-model byte (the colour-variant selector) and the codename the
+/// depot match falls back on. Models that collapse to one key would resolve
+/// to the same depot files anyway.
+pub(crate) fn model_key((model, codename): &(DeviceModelInfo, Option<String>)) -> String {
+    format!(
+        "{:02x}:{:04x}:{:04x}:{:04x}:{}",
+        model.extended_model_id,
+        model.model_ids[0],
+        model.model_ids[1],
+        model.model_ids[2],
+        codename.as_deref().unwrap_or_default()
+    )
+}
+
+/// A manual asset action requested from the Settings → Assets tab, pushed to
+/// the main event loop via [`AssetControl`].
+pub enum AssetCommand {
+    /// Force-fetch assets for the connected devices now, bypassing the
+    /// automatic download policy.
+    Refresh,
+    /// Delete the per-user cache, then re-fetch.
+    ClearCache,
+}
+
+/// Global handle the Settings window uses to push [`AssetCommand`]s into the
+/// main loop, mirroring how the Add Device window drives pairing.
+pub struct AssetControl(pub tokio::sync::mpsc::UnboundedSender<AssetCommand>);
+
+impl gpui::Global for AssetControl {}
+
+/// Minimum gap before re-attempting a failed sync, doubling with each
+/// consecutive attempt and capped at a minute. The first attempt is
+/// immediate (`last_sync_at` is `None`); after that a permanently-down host
+/// is polled ever more slowly (1s, 2s, 4s … 60s) instead of on every tick,
+/// while a recovered host still self-heals on the next attempt.
+pub(crate) fn sync_retry_delay(attempts: u32) -> Duration {
+    ExponentialBuilder::default()
+        .without_max_times()
+        .build()
+        .nth(attempts.saturating_sub(1).min(6) as usize)
+        .unwrap_or(Duration::from_mins(1))
+}
+
+/// Refresh the asset cache: the shared index always, plus the depots for
+/// `models`. Returns `true` when the sync completed and `false` when it
+/// failed and should be retried. Runs on a dedicated background thread —
+/// the HTTP layer's blocking retries are fine here. (Whether sync runs at
+/// all is the caller's gate: the automatic path checks `should_run` once at
+/// startup plus the auto-download setting; the Settings → Assets manual
+/// actions always fetch, even in a release build that would otherwise serve
+/// only bundled art.)
+pub(crate) fn run_asset_sync(models: &[(DeviceModelInfo, Option<String>)]) -> bool {
+    let server = std::env::var("OPENLOGI_ASSETS").unwrap_or_else(|_| DEFAULT_BASE.to_string());
+    match sync(&server, models) {
+        Ok(()) => true,
+        Err(e) => {
+            warn!(error = ?e, "asset sync failed — will retry with backoff");
+            false
+        }
+    }
+}
+
+/// Flatten every paired device's HID++ model snapshot — that's what the
+/// asset sync feeds into the registry lookup.
+pub(crate) fn collect_models(
+    inventories: &[DeviceInventory],
+) -> Vec<(DeviceModelInfo, Option<String>)> {
+    inventories
+        .iter()
+        .flat_map(|inv| inv.paired.iter())
+        .filter_map(|p| p.model_info.clone().map(|m| (m, p.codename.clone())))
+        .collect()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::sync_retry_delay;
+    use std::time::Duration;
+
+    #[test]
+    fn retry_delay_doubles_then_caps() {
+        assert_eq!(sync_retry_delay(1), Duration::from_secs(1));
+        assert_eq!(sync_retry_delay(2), Duration::from_secs(2));
+        assert_eq!(sync_retry_delay(3), Duration::from_secs(4));
+        assert_eq!(sync_retry_delay(5), Duration::from_secs(16));
+        // Caps at 60s and never overflows the shift for large attempt counts.
+        assert_eq!(sync_retry_delay(7), Duration::from_mins(1));
+        assert_eq!(sync_retry_delay(u32::MAX), Duration::from_mins(1));
+    }
 }
